// src/app/model/forecast-hours-detailed.ts
export class ForecastHoursDetailed {

  hour: number;
  icon: string;
  condition: string;
  temp: number;
  dewPoint: number;
  windChill: number;
  relativeHumidity: number;
  atmosphericPressure: number;
  precipitation: number;
  windSpeed: number;
  windGustSpeed: number;
  windDirectionDegree: number;
  windDirectionCardinal: string;

  constructor() {
  }
}